@@ -8,15 +8,16 @@
     "@angular/compiler": "^10.0.9",
     "@angular/core": "^10.0.9",
     "@angular/forms": "^10.0.9",
-    "@angular/localize": "^9.1.12",
+    "@angular/localize": "^10.2.0",
     "@angular/platform-browser": "^10.0.9",
     "@angular/platform-browser-dynamic": "^10.0.9",
     "@angular/router": "^10.0.9",
+    "@ngxs/store": "^3.7.0",
     "bootstrap": "^4.5.2",
     "jquery": "1.9.1 - 3",
     "popper.js": "^1.16.1",
     "rxjs": "^6.6.2",
-    "tslib": "^2.0.1",
+    "tslib": "^2.0.0",
     "zone.js": "^0.10.3"
   },
   "scripts": {
@@ -29,23 +30,23 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^0.1001.7",
-    "@angular/cli": "9.0.0-rc.7",
-    "@angular/compiler-cli": "9.0.0-rc.7",
-    "@angular/language-service": "~7.0.0",
+    "@angular/cli": "10.1.7",
+    "@angular/compiler-cli": "10.2.0",
+    "@angular/language-service": "~10.2.0",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
+    "@types/node": "^8.10.65",
     "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.0",
+    "jasmine-core": "~3.5.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.0.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "protractor": "~7.0.0",
     "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.1.1"
+    "tslint": "~6.1.0",
+    "typescript": "~4.0.3"
   }
 }
