@@ -15,10 +15,26 @@ import { HttpServiceService } from './services/http-service.service';
 import { HttpClientModule } from '@angular/common/http';
 import { ProductListComponent } from './components/product-list/product-list.component';
 import { SearchBoxComponent } from './components/search-box/search-box.component';
+import { NgxsModule } from '@ngxs/store';
+import { ShoppingCartState } from './models/states/ShoppingCart.state';
+import { ShoppingCartComponent } from './components/shopping-cart/shopping-cart.component';
+
+
+
 
 @NgModule({
-  imports:   [BrowserModule, FormsModule, ReactiveFormsModule, CommonModule, CustomValidatorModule, BrowserModule, HttpClientModule],
-  declarations: [ AppComponent, HeaderComponent, FooterComponent, CompteFormComponent, RecapComponent, MatchDirective,  FormatPhonePipe, ProductListComponent, SearchBoxComponent],
+  imports:   [
+      NgxsModule.forRoot([
+          ShoppingCartState
+      ]),
+      BrowserModule, 
+      FormsModule, 
+      ReactiveFormsModule, 
+      CommonModule, 
+      CustomValidatorModule, 
+      BrowserModule, 
+      HttpClientModule],
+  declarations: [ AppComponent, HeaderComponent, FooterComponent, CompteFormComponent, RecapComponent, MatchDirective,  FormatPhonePipe, ProductListComponent, SearchBoxComponent, ShoppingCartComponent],
   bootstrap:    [ AppComponent ],
   providers: [HttpServiceService]
 })
