@@ -3,6 +3,9 @@ import { Observable } from "rxjs/internal/Observable";
 import { Product } from "../../models/Product";
 import { HttpServiceService } from "../../services/http-service.service";
 import { of } from "rxjs";
+import { Store } from '@ngxs/store';
+import { AddProduct } from 'src/app/models/actions/ShoppingCart.action';
+import { Article } from 'src/app/models/ShoppingCart';
 
 @Component({
   selector: "app-product-list",
@@ -14,7 +17,7 @@ export class ProductListComponent implements OnInit {
 
   displayProducts: Observable<Product[]>;
 
-  constructor(private httpService: HttpServiceService) {}
+  constructor(private httpService: HttpServiceService, private store : Store) {}
 
   ngOnInit() {
     this.productsObservable = this.httpService.getProductData();
@@ -25,5 +28,10 @@ export class ProductListComponent implements OnInit {
 
   onFilteredData(event: Product[]): void {
     this.displayProducts = of(event);
+  }
+
+  addProductToShopppingCart(product: Product) : void{
+    let article : Article = Article.fromProduct(product);
+    this.store.dispatch(new AddProduct(article));
   }
 }
